UX update
Added fade in and fade out on most of the levels.

The Major reasoning behind this is designing it so that a player when
they first enter are not intially blinded by the lights and their eyes
can adjust.

Main Menu screen is going to be worked on to be a bit more juicier to
have a bit smoother animations for the player to select and a noise when
selected. (this way it is more interesting and gives feedback to the
player).

Will also be updating the menu systems to incorporate any matinees that
we have.